operator: checksum algorithm use resourceVersion if generation null
Some Referent types, such as ConfigMap and Secret, do not have a
metadata.generation. I think this is because the state of the resource
is entirely contained in the resource, there is nothing to manifest.
That's why they use `data` as the top level property rather than `spec`
and don't need a status. We can instead use the resourceVersion in the
checksum for these types.

Signed-off-by: Robert Young <robeyoun@redhat.com>