operator: annotate KafkaProtocolFilter with referent checksum
If a KafkaProtocolFilter configuration contains any referent secrets and/or
configmaps, then we checksum them all and set the
`kroxylicious.io/referent-checksum` annotation on the
KafkaProtocolFilter metadata. This is updated whenever the referents are
modified. It considers the referent uid and resourceVersion.

The operator cluster role needed additional permissions to patch the
annotations onto the KafkaProtocolFilter resource, where it only had
those permissions for the status subresource before.

Why:
To redeploy the proxy on any referent change we want each resource to be
be annotated with a checksum of it's referents. This checksum will be
used in future to redeploy the proxy on any change.

Signed-off-by: Robert Young <robeyoun@redhat.com>