operator: add referent checksum to KafkaService
If a KafkaService resource has a referent certificate secret, or a
referent trustAnchor configmap, the KafkaServiceReconciler will add a
kroxylicious.io/referent-checksum annotation to the KafkaService's
metadata.

Why:
We intend to consume this annotation in aggregating reconcilers to drive
rolling of the Proxy pods when any config/secret data changes. The proxy
has no ability to reload configuration currently so we are making the
operator handle rolling it.

Signed-off-by: Robert Young <robeyoun@redhat.com>