Refactor reconcilers to operate on ClusterResolutionResult
Why:
Previously the ProxyResolutionResult contained referent maps and the
client code had to dereference them again. It seemed ugly to have to go
back in and out of the API when the DependencyResolver was already
dereferencing everything. Why should the clients have to do it again?

This extends ClusterResolutionResult to contains a full graph of the attempted
resolutions, including the Referent resource if we could obtain it. We
need to expose this partial result to generate the ingresses, to try and keep
port allocation as stable as we can (maybe we should drop this idea? we
decided to limit it to one clusterIp service and will have a fixed port
for SNI purposes).

The end result is the proxy reconciler does not have to dereference the
resources again, it can operate on the ClusterResolutionResult

We are about to add functionality to the VirtualKafkaCluster
reconciler where we will need to access all referents to compute a
checksum from their metadata. Until now we haven't needed to access the
referents, so they aren't part of the ClusterResolutionResult, this
structure will also make it easy to obtain all the Referents for
checksumming in a subsequent PR.

Signed-off-by: Robert Young <robeyoun@redhat.com>