operator: enable clusterIP ingress to use SNI
Adds an optional spec.clusterIP.nodeIdentifiedBy configuration to KafkaProxyIngress.
Possible values are 'port' or 'SNI', default is 'port'. If 'SNI' is
selected, then the `spec.clusterIP.protocol` must be 'TLS' (enforced by
apiserver using CEL).

When 'SNI' is selected, the operator will manifest:
* a bootstrap k8s Service named `${virtualClusterName}-${kafkaProxyIngressName}`. Note
that this is identical to the service name used by 'port'. This is
important because the VirtualKafkaClusterReconciler knows how to get
this Service by name to obtain the bootstrap servers to put in the VKC
status for the ingress.
* a k8s Service per nodeId named `${virtualClusterName}-${kafkaProxyIngressName}-${nodeId}`.
* The proxy configuration gateway for virtual clusters using this
ingress will be configured with bootstrapAddress like:
`${virtualClusterName}-${kafkaProxyIngressName}.${proxyNamespace}.svc.cluster.local`
and an advertisedBrokerAddressPattern like:
`${virtualClusterName}-${kafkaProxyIngressName}-$(nodeId).${proxyNamespace}.svc.cluster.local`
* a container port on the proxy 9291 (shared by all such ingresses)

For example if the VirtualKafkaCluster is named my-cluster and the
KafkaProxyIngress is named my-ingress and there are 3 upstream nodes
with ids (0, 1, 2), then we will manifest services:
* my-cluster-my-ingress port 9291 targetPort 9291
* my-cluster-my-ingress-0 port 9291 targetPort 9291
* my-cluster-my-ingress-1 port 9291 targetPort 9291
* my-cluster-my-ingress-2 port 9291 targetPort 9291
And configure the Proxy gateway with:
- bootstrapAddress: my-cluster-my-ingress.proxy-ns.svc.cluster.local:9291
- advertisedBrokerAddressPattern: my-cluster-my-ingress-$(nodeId).proxy-ns.svc.cluster.local

Why:
Users may wish to use SNI rather than portPerBroker as it's a safer
mechanism. We can allow many ingresses to use SNI on the same proxy
without risking mis-identification of a client connection.

This keeps the VKC status.ingresses reporting of bootstrap servers
working. It assumes there will be a Service named
`${clusterName}-${ingressName}` and that the bootstrap port is the first
port of the service. So the path of least resistance is to have the bootstrap
service be named the same across port per broker and SNI.

This does mean that we have knowledge of the Ingress implementation over
in the VKC reconciler and may have to evolve this as we add further ingress
types.

Signed-off-by: Robert Young <robeyoun@redhat.com>