bugfix: operator do not emit empty downstream trust object
Fixes #2176

If you supply downstream TLS server certs but no trust, the operator was
manifesting it as an empty object in the configuration YAML. This cannot
be decoded using jackson deduction. For now lets change it to not set
trust if it isn't supplied and we can improved the configuration API
separately.

Signed-off-by: Robert Young <robeyoun@redhat.com>