Support $(virtualClusterName) in SNI advertisedBrokerAddressPattern
If $(virtualClusterName) is present in the advertisedBrokerAddressPattern
(or deprecated brokerAddressPattern), then it is replaced with the name
of the gateway's VirtualCluster.

So if the advertisedBrokerAddressPattern is
`$(virtualClusterName)-broker-$(nodeId).mycluster.com:443` and the
virtual cluster name is my-cluster, then the advertised address for node 3
will be `my-cluster-broker-3.mycluster.com:443`. The proxy will map
`my-cluster-broker-3.mycluster.com` to node 3 for my-cluster when it
receives a connection.

Why:
We want users of the operator to be able to define a single Ingress that
defines the bootstrap address with a cluster name placeholder so that we
can inject it into multiple virtual cluster's gateway configs. The
configuration will be identical and not need the operator to know how to
construct different configurations for each cluster.

Signed-off-by: Robert Young <robeyoun@redhat.com>