Support $(virtualClusterName) in SNI bootstrap address
If $(virtualClusterName) is present in the SNI bootstrapAddress, then it
is replaced with the name of the gateway's VirtualCluster.

So if the bootstrapAddress is `$(virtualClusterName)-bootstrap.mycluster.com:9292`
and the virtual cluster name is `my-cluster`, then the Proxy will map the SNI host
`my-cluster-bootstrap.mycluster.com` to this gateway.

Why:
We want users of the operator to be able to define a single Ingress that
defines the bootstrap address with a cluster name placeholder so that we
can inject it into multiple virtual cluster's gateway configs. The
configuration will be identical and not need the operator to know how to
construct different configurations for each cluster.

Signed-off-by: Robert Young <robeyoun@redhat.com>