Require VirtualClusters to be unique case-insensitively
We want to enforce the property that virtual cluster names are
valid DNS labels so that we can more safely inject it into broker
addresses in future. We are working towards the Operator being able to
provide a single SNI gateway configuration using cluster name placeholders
injected into multiple virtual clusters.

If we allowed two clusters to have the same name with case differences
they would produce the same DNS label (case insensitive as far as DNS
cares) and we may produce the same addresses for two clusters. Instead
we require the names to be unique, case insensitively.

Signed-off-by: Robert Young <robeyoun@redhat.com>