operator: require cluster name plus ingress name to be <= 51 chars
Why:
When constructing a clusterIP ingress we currently use the vkc name plus
the kpi name with a hyphen between them. A k8s service name must be 63
characters or less. We want to reserve 11 characters to allow the
possibility for injecting node ids up to integer max for clusterIP SNI
access, leaving 52 characters available for the names (and a hyphen).

Signed-off-by: Robert Young <robeyoun@redhat.com>