operator: only consider fresh referent ResolvedRefs statuses
When we are observing Referent resources to determine if they have a
status condition declaring that that referent has unresolved
dependencies, we should only consider fresh statuses where the
observedGeneration of the condition is the same as the
metadata.generation of the referent. A ResolvedRefs condition for an
older generation is not material to the decision as it doesn't reflect
the current spec.

Note that we have other means to prevent the aggregating reconcilers
processing referents with stale statuses, this covers the case where the
status is fresh, but it contains an old condition for whatever reason.

Signed-off-by: Robert Young <robeyoun@redhat.com>