operator: refactor ingress model
This is preparing for adding a second ingress type.

* moves downstream TLS to IngressModel, since TLS is under the clusterIP
node in the config tree, and we will be introducing a loadbalancer
node at the same level we need this to prevent the reconciler having to
know how to retrieve downstream TLS details for each ingress type.
Instead it can ask for the downstream TLS and the specific ingress model
can handle the details.
* renames IngressInstance to IngressModel to match the other model class
  naming for this logical layer.
* inlines IngressDefinitionBuilder into IngressAllocator, this was the
  only client calling a couple of static methods.
* renames IngressModel to IngressModelResult, to match our other naming
  for result classes
* Makes IngressModel responsible for supplying a NodeIdentificationStrategy
so that the new ingress type will be able to supply an SNI strategy
* Introduces a static bootstrapServiceName method on ClusterIPIngressDefinition
to clarify the link between the service name and the VKC ingress status,
where we are loading the Service as a secondary resource by name.
* Makes the clusterIP implementation of IngressModel private, reducing
  visibility
* Reduces visibility of IngressDefinition, as it's just a means to an
  ends of obtaining an IngressModel once the ports are allocated.

Signed-off-by: Robert Young <robeyoun@redhat.com>