operator: refactor ingress model
* renames ProxyIngressModel to ProxyNetworkingModel
* renames VirtualClusterIngressModel to ClusterNetworkingModel
* renames IngressInstance to ClusterIngressNetworkingModels
* renames IngressAllocator to NetworkingPlanner, which plans a
ProxyNetworkingModel of resources and configuration to be manifested
to enable a client to connect to a virtual cluster via a KafkaProxyIngress
* moves downstream TLS to ClusterIngressNetworkingModel, this saves having to
re-correlate the ingress back to a TLS blob pulled from the VKC ingresses.
We can associate it as part of networking model planning.
* inlines IngressDefinitionBuilder into NetworkingPlanner, this was the
  only client calling a couple of static methods.
* Makes ClusterIngressNetworkingModel responsible for supplying a
NodeIdentificationStrategy so that the new ingress type will be able to
supply an SNI strategy
* Introduces a static bootstrapServiceName method on ClusterIPClusterIngressNetworkingModel
to clarify the link between the service name and the VKC ingress status,
where we are loading the Service as a secondary resource by name.
* Inlines IngressDefinition into NetworkPlanner, as it's just a means to an
ends of obtaining a ClusterIngressNetworkingModel, once the ports are allocated,
it should be private concern.

Why:
Ingress is a specific kubernetes term and it creates some confusion. The
networking model is about planning what k8s resources should be
manifested for clients to connect to, such as Services and Routes, but
it is also about the backend plumbing required to connect those services
to the Proxy. That is the container ports and corresponding proxy
configuration that need to be wired into the proxy. The ports, hostnames
and downstream TLS have to marry up across the k8s networking resource,
container resources and proxy configuration. We call this networking.

This is preparing for adding a second ingress type.

Signed-off-by: Robert Young <robeyoun@redhat.com>
Co-authored-by: Sam Barker <sam@quadrocket.co.uk>