Implement LoadBalancer ingress
If any referenced KafkaProxyIngress contains a loadBalancer property, we:
1. create a single Service with type LoadBalancer for the proxy. This is
   to be shared by all LoadBalancer ingresses.
2. add a shared-sni container port to the proxy, 9291. This is the port
   all of the shared Service ports will target.

Then, for each loadbalancer KafkaProxyIngress referred to by a VKC we
add a gateway to that VKC's proxy config. The gateway uses SNI and we
pass through:
- the bootstrapAddress from the KPI custom resource, with the shared-sni
  container port appended, as this is the address the proxy will listen
  on.
- the advertisedBrokerAddressPattern from the KPI custom resource, with
  the client-facint Service port appended (currently defaulted to 9083)
  as this is the port clients will have to connect to.

In future the port exposed by the shared Service for a KafkaProxyIngress
will be customizable, the intent is that the shared Service will
accumulate all the client-facing ports across all ingresses and map them
all to the shared-sni port on the proxy container.

We have not implemented VKC ingress status for loadbalancer ingresses
yet.

Signed-off-by: Robert Young <robeyoun@redhat.com>
Co-authored-by: Keith Wall <kwall@apache.org>
Signed-off-by: Ujjwal <ukalia@confluent.io>