Annotate Services with the bootstrap servers that they host
ClusterIP ingresses manifest a Service that hosts the bootstrap port.
Loadbalancer ingresses use a shared Service that will host the defined
bootstrap for those ingresses.

We want a way to report in the VKC reconciler on the bootstrap servers
for each ingress in a VKC ingresses array. Currently we do that by
looking up the ClusterIP bootstrap service by name and then construct
the bootstrap servers. We want to present the same information for
LoadBalancer ingresses, but we cannot reconstitute the bootstrap servers
just from the Service name and namespace. We would need a different
mechanism to recompute the bootstrap servers.

Instead, we can require all Services that will be used for bootstrapping
to be annotated with the expected bootstrap servers that will be used to
connect to them. This provides a uniform way for LoadBalancer and
ClusterIP to present their bootstrap servers to the VKC reconciler.

This requires the VKC to recompute the bootstrap servers

Signed-off-by: Robert Young <robeyoun@redhat.com>