Refactor: consolidate Annotation read/write into a single class
This hides the referent checksum annotation implementation from all the places that
need to read/write it. This is a positive because it means that if we
needed to encode more data in future, we could migrate transparently to
a new key/value, while continuing to support reading the old format. Also we can
share some annotation read/write logic with the new bootstrap annotations code.

Signed-off-by: Robert Young <robeyoun@redhat.com>