Refactor how we describe shared loadbalancer requirements
Why:
Previously we used the existence of a shared SNI port requirement to
imply that we also want to collect the bootstrap servers to be annotated
on. Now we instead have an optional
SharedLoadBalancerServiceRequirements interface that declares which
client facing ports should be exposed, and also declares what bootstrap
servers need to be annotated into the Service metadata.

Signed-off-by: Robert Young <robeyoun@redhat.com>