Introduce KafkaProxy.spec.infrastructure.labels
We want to introduce an infrastructure object as a better expression of
the intent, then nest more specific infrastructure manipulations under
it, for example we could use 'infrastructure.proxyPod` for a template
for the proxy pod.

`infrastructure.labels` is intended to be applied to all dependant
resources eventually, though initially it is just applied to the proxy
pod, replacing `KafkaProxy.spec.podTemplate`.

Signed-off-by: Robert Young <robeyoun@redhat.com>