Enable user to specify proxy pod resource limits and requests
Adds `spec.infrastructure.proxyPod` to KafkaProxy. This is intended to have a
structure matching a pod template spec, restricted to just the fields we
wish users to have control over.

By default we give the dependant Proxy pod 500m cpu and 512Mi memory for
both requests and limits.

If the user supplies any resources via the KafkaProxy infrastructure,
then those resources are applied to the deployment. We do not attempt to
combine them with the default limits/requests. This enables the user to
supply an empty `resources` object if they wanted to remove limits/requests entirely.

Why:
Users need to be able to control these values, tailoring it to their
Filters and environment.

Signed-off-by: Robert Young <robeyoun@redhat.com>