CRD: add resources controls to KafkaProxy
1. Adds `spec.infrastructure` as a parent for anything concerned with
   configuring implementation-specific pieces of infrastructure at a platform level.
2. Adds `spec.infrastructure.proxyContainer` as a parent for
   configuration specific to the proxy container.
3. Adds `spec.infrastructure.proxyContainer.resources` for configuring
   the container requests/limits as you would in a PodSpec, like:
   ``` yaml
   resources:
     requests:
       cpu: '500m'
       memory: '512Mi'
     limits:
       cpu: '1000m'
       memory: '1024Mi'
   ```

We have intentionally tried to limit the proxyContainer spec to something
minimal, rather than trying to replicate a full PodSpec. We think
it might be more flexible in future to introduce something like
`spec.infrastructure.proxyPodSpecRef` to reference a full PodSpec to
give the user more control over the details of the pod spec without us
having to duplicate the spec. Perhaps with a mechanism to generate
an initial PodSpec resource.

Signed-off-by: Robert Young <robeyoun@redhat.com>