Manifest KafkaProxy.spec.infrastructure.proxyContainer.resources
If the user has supplied resource requests/limits for the proxy
container via KafkaProxy.spec.infrastructure, then these resources are
set on the proxy container.

Note that we do not attempt to merge the user supplied resources with
the defaults in any way. This is to enable users to do things like
specify requests without limits. The user supplied resources are applied
verbatim.

This gives users control over the cpu/memory limits/requests of their
proxy containers.

Signed-off-by: Robert Young <robeyoun@redhat.com>