Run 3 parallel forks of kroxylicious integration-tests
We currently have to tell the proxy what ports to listen on. The
integration tests all rely on stopping and starting the proxy on
that port, and make assumptions that certain other ports are available
to bind to if they wish to create further virtual clusters.

This change configures things so that surefire runs the tests in 3
parallel forks and passes the forkNumber as a system property to the
fork, the tests then use this to offset the ports. Each fork can use
50 ports without colliding.

Signed-off-by: Robert Young <robeyoun@redhat.com>