fix: stabilize sporadically failing system test (#2399)
* fix: stabilize sporadically failing systemm test

why: we were sporadically seeing failing tests where the error message was a conflicting job name.  now we delete
the delete in the case where the job/pod run is successful.  in the case where it is not successful, we retain
job/pod so that it will be captured as part of the diagnostic dump.

Signed-off-by: Keith Wall <kwall@apache.org>