Add hostname/ip information to VKC loadbalancer status
When reporting the VKC status for an ingress:

If the ingress is associated with a Service with a `loadBalancer` object in its
status we report each `ingress` point's ip or hostname in the VKC status.

Why:

The first time a user sets up a LoadBalancer KafkaProxyIngress they will
be responsible for configuring their DNS to resolve some hostnames to
the loadbalancer ip/hostname. That means we must instruct users on how
to locate the shared SNI service and extract its details.

Signed-off-by: Robert Young <robeyoun@redhat.com>
Co-authored-by: Sam Barker <sam@quadrocket.co.uk>