Build and deploy website
We use the Dockerfile from the website repository to build the website,
including the documentation from the current version. We build on main
push, or can be triggered by a workflow_dispatch with branch as
paremeter.

Then we deploy the output using a deploy pages action.

We override the baseurl to kroxylicious to match this repositories name
as the repository name should make up part of the Pages url, like
https://kroxylicious.io/kroxylicious/ Jekyll must be told about the
baseurl so that generated links can include it. (note there are many
broken links that need to be fixed to consider baseurl)

It is possible for the repository to override Jekyll configuration, so
that we can set the site.url or baseurl, which is useful for deploying
on a fork.

Signed-off-by: Robert Young <robeyoun@redhat.com>