Fail fast on unknown properties
Why:
The proxy was lenient, allowing you to start up with unknown properties
in the configuration file. This can have a couple of bad effects:

1. users may see the `bootstrapServers` configuration of `targetCluster`
   and guess that they can plug in other typical kafka client
configurations. This would be quite common in other kafka integrated
technology. The proxy then starts up happily despite their config not
affecting the proxy at all.
2. a typo in an optional field name may cause the proxy to happily start
   but not apply the user's behaviour. For example a typo in a gateway `tls`
field might cause no TLS settings to be applied, and the proxy server
would not protect that gateway's endpoints.

There is a downside to being harsh in that we lose forward
compatibility. If we ever implemented dynamic configuration reloading,
forward compatibility could be a useful property during deployments of
new versions. Old proxies could be exposed to new configuration versions
with new fields, and potentially be able to keep operating.

However I think the real problems we are battling right now with users
getting burned by lenient config are problematic.

Signed-off-by: Robert Young <robeyoun@redhat.com>