fix: fix flakey KafkaClientTest (#2453)
why: test instability.  we saw a

javax.net.ssl.SSLHandshakeException: No subject alternative names matching IP address 127.0.0.1 found

during a test run.   The test configures the certificate to expect localhost then the client connects
with 127.0.0.1.  There is no reason for the test not to connect with the symbolic name (and avoid any need
for a rDNS).

Signed-off-by: Keith Wall <kwall@apache.org>