bugfix: docker build not triggering for pull requests from forks
Refactors to prevent accessing vars and secrets when the action is
triggered by pull request.

Why:

the top level if was preventing PRs from triggering the docker build
because they can't access the repository vars/secrets for security
reasons. When building a PR we just want to excercise the build and
don't need to push anywhere, so we can use an arbitrary tag.

Signed-off-by: Robert Young <robeyoun@redhat.com>