bugfix: forward v0 API_VERSIONS bytes on upstream unknown error response
If we receive an API_VERSIONS response from upstream with error code 35
(UNKNOWN_VERSION), following KIP-511 we assume that the response was
written with api version 0. This is the response prescribed for when
the server recieves an API_VERSION api version it does not understand.

A user discovered that a Kafka client (twmb/franz-go) was failing to
connect to Kafka 3.8.0 via the proxy. The flow is something like this:
1. franz-go sends a version 4 API_VERSIONS request
2. kroxylicious 0.13.0 understands version 4 and forwards it upstream
3. upstream doesn't understand and responds with v0 and error code 35
4. kroxylicious tries to decode it, we leniently tried to decode it as
   v4, then retried with a v0 decode on exception.
5. kroxylicious assumed the frame was v4 and forwarded it to the client,
   encoded at v4.
6. franz-go expects that error code 35 implies the remainder of the
   message is written with v0 and it barfs because there are too many
   bytes in the response.

Note that we are depending on the KIP-511 contract and not checking that
the upstream really responded with v0 bytes.

Signed-off-by: Robert Young <robeyoun@redhat.com>