fail fast if upstream Api Versions is err 35 but not v0
If the upstream is responding to an API_VERSIONS request with an
UNKNOWN_VERSION error code, then we expect it to be a v0 response body
per KIP-511. The proxy decode logic is lenient, attempting to decode
with the request api version, then trying again with v0 on error.
If we encountered an UNKNOWN_VERSION error but successfully decoded it
at a higher api version, we are encountering an incorrect upstream
behaviour and should fail fast.

Signed-off-by: Robert Young <robeyoun@redhat.com>