Test KafkaClient ensures frame entirely read by header/body decode
We want to use KafkaClient to declare what response apiVersion we expect
the proxy to respond with. The kafka-client code we depend on to read
the messages will sometimes be backwards compatible, for example if we
try to read a v1 or v2 ApiVersions response message, then it will
successfully read the v0 portion of the message, leaving some additional
fields like throttleTimeMs unread.

By KIP-511 the contract is we will send a v0 response, not a v0
compatible response, so I think it's good for the test client to require
that all bytes are read at the response apiVersion we declare we expect.

Signed-off-by: Robert Young <robeyoun@redhat.com>