Split integration tests into their own action
Why:
The unit+integration test branch currently takes ~33 minutes with 24
minutes dedicated to the integration tests step.

Splitting it into a branch for the UTs and linting and another for the
integration tests brings the runtimes to ~22 minutes for ITs and ~11
minutes for the UTs and linting. This should reduce the build time by
a third as the IT branch will still be the longest by a couple of
minutes over the operator build.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>