Add a linting workflow and trigger PR sonar build after it
Adds an action to quickly check that the proxy build compiles and lints.
The intent is to trigger the sonar build sooner.

Why:
We currently trigger the sonar run after the operator build (~16 minutes)
the sonar workflow_run takes ~11 minutes. So time to report sonar on a
PR is ~27 minutes.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>