refactor(docs): update documentation versioning attributes
Replace gitRef with KroxyliciousGitRef, this is used when we want to
link to source code in github. At development time it will be 'main' and
at deploy time it will be the release tag, e.g '0.14.0'.

Introduce KroxyliciousReleaseTag, this is used when we want to link to
release artefacts. At development time there is nothing to point at so
it can always be v{KroxyliciousVersion} to give some clue that we will
be pointing at the right artefact.

Refactors a couple of usages to be in terms of some new attributes that
use these two attributes to create links to the github source tree, blob
and raw scripts.

Also the release scripting is updated to set them appropriately. At
release time the KroxyliciousVersion should be set to the new release
version and the KroxyliciousGitRef should be set to the release tag. Then
the development snapshot commit should set KroxyliciousVersion to the
next version and the KroxyliciousGitRef should be set to 'main'.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>