add experimental configuration options to control encryption buffer
Users max wish to control the buffer sizes to match their upstream
brokers max message size a well as tune the initial buffer size to match
their workload.

This is spurred by issue #2504 where a large message cannot be
encrypted. I chose to make the settings experimental as the issue is
prompting rethinking of the encryption mechanisms.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>