Ensure that we will attempt to encrypt with the max buffer size
Now that the minimum and maximum buffer sizes are configurable the
minimum being doubled may produce a size greater than the max. It would
be a shame to throw at this point without ever trying the encrypt at the
maximum configured buffer size.

For example with min size 100 and max size 150, we would previously
never attempt at 150 because our second attempt would have doubled the
buffer size to 200 and thrown because it exceeds max.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>