generate release manifest file and download page for website
Why:
At release time we need to create (for version X.Y.Z):
1. _data/release/X_Y_Z.yaml file describing the release assets
2. download/X.Y.Z/index.md file, boilerplate wiring up the download page

Note: we currently do not have the released image SHAs at documentation
PR creation time. So the manifest in the created docs PR will contain
placeholders to be manually updated (blergh!) for now. The PR body
is updated with a reminder, and there is automation in the website repo
requiring them to be replaced before the PR can be merged.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>