feat(test): Add resilience tests for bootstrapping logic
* Add resilience tests for bootstrapping logic

We want to prove that we can bootstrap if only a subset of the target
bootstrap servers are available.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>

* Add resilience tests covering bootstrapping on proxy restart

The proxy has special logic to handle the case where active clients want
to immediately connect to a broker port before the proxy is aware of the
upstream address for that broker. We hijack the connection, and connect
to an upstream bootstrap server to obtain the Metadata containing the
upstream addresses. Then we close the connection. From this point we can
correctly route broker-port connections to the relevant upstream broker.

We need to ensure this logic is resilient too if the first bootstrap
server is offline.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>

---------

Signed-off-by: Robert Young <robertyoungnz@gmail.com>