Bootstrap bindings are no longer tied to a fixed upstream.
They weill now reselect on invocation. This does have the implication that different instances of the proxy will bootstrap to different brokers. Which has some interesting edge cases in the event of a split brain situation where different proxy instances can discover different metadata - however this was always the case in k8s as the bootstrap service was doing round-robin anyway.

Signed-off-by: Sam Barker <sam@quadrocket.co.uk>