fix: sendRequest and short-circuit response do not compose
Transfers the promise and recipient internal fields when a Filter
generates a short-circuit response to an out-of-band request initiated
by a downstream filter.

Previously, the internal fields were not transferred, breaking the
internal request mechanism as the promise was lost forever. It caused
the connection to hang and never respond to the client.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>