Add an Azure Key Vault KMS implementation
We support pluggable KMS implementations for the Record Encryption
Filter. This adds support for Azure Key Vault and it's related variant
Managed HSM.

Features:
1. Support 'RSA' and 'RSA-HSM' key types, wrapping using `RSA-OAEP-256`
2. Support 'oct' and 'oct-HSM' key type with AES-GCM wrapping (will only
   work with Managed HSM)
3. Support only client credentials oauth flow with Entra using clientId
   + clientSecret. This supports workloads running anywhere.
4. Support TLS customization of the authentication client and key vault client.
5. DEK bytes are generated proxy-side with a SecureRandom
6. Edek stores the keyName, keyVersion, edek.
7. User will supply tenantId for authentication. It is possible to
   obtain this by making an HTTP request to Key Vault with no
   Authorization header. But it should be an easy thing for users to
   supply.

Auth:
With the oauth client credentials flow we obtain a Bearer token along
with a remaining lifetime in seconds (default 1 hour). We share the
token across all filters for a single Filter definition. When it gets
near to expiry (1 minute away), then we trigger an asynchronous refresh.
The established token continues to be used while the refresh is in
progress.

Example configuration:
```yaml
kms: AzureKeyVault
kmsConfig:
  keyVaultBaseUrl: https://kv-my-vault-name.vault.azure.net
  tls:
    ... client trust for key vault
  entraIdentity:
     oauthEndpointUrl: https://login.microsoftonline.com // optional defaults to this value
     clientId:
       passwordFile: /path/to/id
     clientSecret:
       passwordFile: /path/to/secret
     tenantId: "abds-1232dsaa"
     scope: https://vault.azure.net/.default // optional, defaults to this value
     tls:
       ... client trust configuration for oauth
```

Excludes com.azure:azure-core-http-netty from azure test support

This is the 1st party implementation of com.azure.core.http.HttpClient
which depends on a stack of netty dependencies. The Lowkey test
container has the ability to return clients configured to use this
client, or to use it's own apach httpclient backed version

The big stack of netty dependencies was a problem in
kroxylicious-integration-tests as it included tcnative-boring-ssl which
altered the TLS behaviour of the proxy under test. So there are risks to
pulling in netty test dependencies as they can alter the proxy
behaviour

We can get away with excluding the netty httpclient as we do are using
the apache httpclient variant only to work with the KMS

Signed-off-by: Robert Young <robertyoungnz@gmail.com>