make KrpcGenerator not modify existing up-to-date sources
We generate to a temporary file and then compare the entire contents to
the existing file in the generated source dir (if it is present). If the
file is identical, we delete the temporary file, leaving the existing
file untouched. Else, we replace it with the new generated file.

Why:
Currently kroxylicious-api will recompile on every build because the
generated sources are overwritten with files with fresh timestamps. This
will provoke all of the dependant modules to recompile. We are
experimenting with re-using the built output in github actions to
parallelize the build via an artifact and currently this behaviour
causes unneccesary rebuilding to occur.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>