Normalize alias, converting underscores to hyphens for Azure KMS
The existing test failed because we use `$(topicName)` as kek selector.
The topic names contain underscores which is not valid as an azure key
name. To workaround I've made the Azure plugin normalize underscores to
hyphens, I figure keeping topicName working in some capacity is useful.
There's an argument the other way that the user is no longer specifying
the kek. Maybe we should introduce another replacement like
`$(underscoreToHyphenTopicName)` or `$(azureNormalizedTopicName)` to
make it explicit. Or we could make it a configuration option on the
plugin so users have to opt in to the behaviour.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>