Exclude com.azure:azure-core-http-netty from azure test support
This is the 1st party implementation of com.azure.core.http.HttpClient
which depends on a stack of netty dependencies. The Lowkey test
container has the ability to return clients configured to use this
client, or to use it's own apach httpclient backed version

The big stack of netty dependencies was a problem in
kroxylicious-integration-tests as it included tcnative-boring-ssl which
altered the TLS behaviour of the proxy under test. So there are risks to
pulling in netty test dependencies as they can alter the proxy
behaviour

We can get away with excluding the netty httpclient as we do are using
the apache httpclient variant only to work with the KMS

Signed-off-by: Robert Young <robertyoungnz@gmail.com>